fix : url 접근시 component 랜더링 후 redirect 오류 해결
main layout 에 loading 상태를 추가하여 login 상태를 판별한 후 component 랜더링으로 변경
